Make jsluice optional via build tags
